Default first capital letter on new subquests. Hide keyboard on tab change. After add new subquest do not hide keaboard and change focus.
