+++ app/src/main/java/io/ipoli/android/app/BaseFragment.java
@@ -1,11 +1,14 @@
 package io.ipoli.android.app;
 
+import android.content.Context;
 import android.os.Bundle;
 import android.support.annotation.Nullable;
 import android.view.LayoutInflater;
 import android.view.MenuItem;
 import android.view.View;
 import android.view.ViewGroup;
+import android.view.WindowManager;
+import android.view.inputmethod.InputMethodManager;
 
 import com.trello.rxlifecycle.components.support.RxFragment;
 
@@ -55,5 +58,17 @@ public abstract class BaseFragment extends RxFragment {
     }
 
     protected void showHelpDialog() {
+    }
+
+    protected void showKeyboard() {
+        getActivity().getWindow().setSoftInputMode(WindowManager.LayoutParams.SOFT_INPUT_STATE_VISIBLE);
+    }
+
+    protected void hideKeyboard() {
+        View view = getActivity().getCurrentFocus();
+        if (view != null) {
+            InputMethodManager imm = (InputMethodManager) getActivity().getSystemService(Context.INPUT_METHOD_SERVICE);
+            imm.hideSoftInputFromWindow(view.getWindowToken(), 0);
+        }
     }
 }
