+++ app/src/main/res/layout/fragment_subquest_list.xml
@@ -52,7 +52,6 @@
                         android:src="@drawable/ic_add_accent_24dp" />
                 </FrameLayout>
 
-
                 <android.support.design.widget.TextInputEditText
                     android:id="@+id/add_subquest"
                     android:layout_width="match_parent"
@@ -61,7 +60,7 @@
                     android:layout_marginEnd="16dp"
                     android:layout_marginStart="72dp"
                     android:imeOptions="actionDone"
-                    android:inputType="text"
+                    android:inputType="textCapSentences"
                     android:maxLines="1"
                     android:text="@string/add_sub_quest"
                     android:textColor="@color/colorAccent"
