+++ app/src/main/res/layout/subquest_list_item.xml
@@ -47,7 +47,7 @@
             android:layout_gravity="center_vertical"
             android:layout_marginEnd="16dp"
             android:layout_marginStart="72dp"
-            android:inputType="text"
+            android:inputType="textCapSentences"
             android:maxLines="1"
             android:textColor="@color/md_dark_text_87"
             android:textSize="@dimen/subquest_item_text_size" />
